fix: stabilize Resource Authorization identity migration tests
They were flaky as we run the migration with different settings with the same
cluster + secondary storage.
This isolated Resource Authorization tests to dedicated test classes.
